requirements.txt
flask
requests
python-dotenv
python-telegram-bot==21.6